requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.27
langchain-groq==0.3.8
langchain-huggingface==0.1.2

faiss-cpu==1.8.0
sentence-transformers==3.0.1

python-dotenv==1.0.1
pydantic==2.9.2

# pip install -r requirements.txt